Reorganize files, modules and prepare tests
