+++ .gitignore
@@ -1,0 +1,6 @@
+### Python ###
+__pycache__/
+.pytest_cache/
+
+### Custom ###
+tests/.out/
