+++ Makefile
@@ -1,0 +1,9 @@
+.PHONY: install test
+
+default: test
+
+install:
+	pip install --upgrade .
+
+test:
+	PYTHONPATH=. pytest
