+++ README.md
@@ -1,1 +1,8 @@
+# GitHub Action to generate test results summary page
 
+## Contribute
+
+### Run tests
+
+1. Install [`pytest`](https://docs.pytest.org/en/6.2.x/getting-started.html#install-pytest)
+2. Run `make test`
